maxregcount set, bug in rating fixed, float -> double need for huge problems

+++ src/ensure.cu
@@ -2,7 +2,7 @@ __device__ int ensure_isnt_nan(struct instance *inst,
 		    	       struct memory *mem) {
 //	return 0;
 	for(int r = 0; r < inst->dim.matrix_height; r++) {
-		float* row = C_ROW(r);
+		double* row = C_ROW(r);
 		for(int c = mem->c_zero; c < mem->c_end; c++) {
 			if(isnan(row[c])) {
 				cont = 0;
@@ -13,7 +13,7 @@ __device__ int ensure_isnt_nan(struct instance *inst,
 	}
 
 	for(int r = 0; r < inst->dim.matrix_height; r++) {
-		float* row = R_ROW(r);
+		double* row = R_ROW(r);
 		for(int c = mem->r_zero1; c < mem->r_end2; c++) {
 			if(isnan(row[c])) {
 				cont = 0;
@@ -24,7 +24,7 @@ __device__ int ensure_isnt_nan(struct instance *inst,
 	}
 
 	for(int r = 0; r < inst->dim.matrix_height; r++) {
-		float* row = P_ROW(r);
+		double* row = P_ROW(r);
 		for(int c = 0; c < inst->dim.parents * inst->width_per_inst; c++) {
 			if(isnan(row[c])) {
 				cont = 0;
@@ -42,7 +42,7 @@ __device__ int ensure_range_child(struct instance *inst,
 //	return 0;
 
 	for(int r = 0; r < inst->dim.matrix_height; r++) {
-		float* row = C_ROW(r);
+		double* row = C_ROW(r);
 		for(int c = mem->c_zero; c < mem->c_end; c++) {
 			if(row[c] > PARENT_MAX) {
 				cont = 0;
@@ -61,7 +61,7 @@ __device__ int ensure_range_parent(struct instance *inst,
 	int count = inst->dim.parents * inst->width_per_inst;
 
 	for(int r = 0; r < inst->dim.matrix_height; r++) {
-		float* row = P_ROW(r);
+		double* row = P_ROW(r);
 		for(int c = 0; c < count; c++) {
 			if(row[c] > 10000) {
 				cont = 0;
@@ -83,8 +83,8 @@ __device__ int ensure_correct_copy(struct instance *inst, struct memory *mem,
 	int cols = inst->num_matrices;
 
 	for(int r = 0; r < row; r++) {
-		float* prow = P_ROW(r);
-		float* crow = C_ROW(r);
+		double* prow = P_ROW(r);
+		double* crow = C_ROW(r);
 
 		for(int c = 0; c < cols; c++) {
 			if(isnan(crow[cstart + c]) || isnan(prow[pstart + c]))
